% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastANCOM.R
\name{fastANCOM}
\alias{fastANCOM}
\title{fastANCOM, a fast method for analysis of composition of microbiomes}
\usage{
fastANCOM(
  Y,
  x,
  Z = NULL,
  rand = NULL,
  zero_cut = 0.9,
  lib_cut = 0,
  struc_zero = FALSE,
  pseudo = 0.5,
  global = F,
  outlier = F,
  sig = 0.05,
  detect.rate = 0.7,
  ref.rate = 0.05,
  effect = T
)
}
\arguments{
\item{Y}{the abundance table,  count/relative or any smoothed data matrix are supported}

\item{x}{the variable of interest}

\item{Z}{the other covariates to be adjusted}

\item{rand}{the random variable of the mixed effect model}

\item{zero_cut}{a numerical fraction between 0 and 1. Taxa with proportion of zeroes greater than zero_cut will be excluded in the analysis. Default is 0.90}

\item{lib_cut}{a numerical threshold for filtering samples based on library sizes. Samples with library sizes less than lib_cut will be excluded in the analysis. Default is 0, i.e. do not filter any sample.}

\item{struc_zero}{whether to detect structural zeros. Default is FALSE}

\item{pseudo}{the pseudo number for zero smoothing, default is 0.05}

\item{global}{whether to perform global test for more than two groups. Default is FALSE}

\item{outlier}{a boolean value, if TRUE the cook's distance is used for outlier detection, the default is FALSE}

\item{sig}{the level of significance to count the reject number, default is 0.05}

\item{detect.rate}{a numerical fraction between 0 and 1. Miorobes with proportion of the reject number greater than it will be identified as associated with the variable of interest, i.e., x, default is 0.7}

\item{ref.rate}{a numerical fraction between 0 and 0.5, the proportion of total microbes with the least reject numbers and these microbes are set as the reference to estimate the effect size in ecosystem, the default is 0.05}

\item{effect}{a boolean value, if TRUE then estimate the effect size based on the reference set, the default is T}
}
\value{
a list with two components:
\itemize{
\item global, if the variable of interest x is continuous or global==FALSE, it returns NULL, otherwise a list with three components
\itemize{
\item marginal, a list contains the estimated coefficient and the covariance structure of the marginal model
\item joint, a list contains 4 elements for the joint model: beta, the coefficients from the joint model; se, the corresponding standard error; statistic, the Wald statistic; and W, the reject number of each component.
\item final, a data.frame contains the 5 columns for the final effect size estimation,
\itemize{
\item log2FC, the adjusted effect size of log-linear model
\item log2FC.SD, the standard errors (SEs) of log2FC
\item log2FC.pval, the p-value of Wald test
\item Reject.number, the reject number of each microbe
\item REJECT, the final result whether each microbe is associated with the variable of interest or not
}
}

\item results, a list contains the results for the pairwise group comparison, each element will be same structure as above
}
}
\description{
fastANCOM fits the mariginal model and then computing the effect for the joint model.
The framework of log-linear regression is used for the log transformed data.
The hypothesis is build with the variable of interest for inference at the ecosystem
level with the framework of ANCOM.
}
\examples{
data <- matrix(rpois(100*60, 10), 60)
b <- rnorm(10, 4)
data[31:60,1:10] <- t(t(data[31:60,1:10])*2^b)
group <- rep(c(0, 1), each=30)
fit <- fastANCOM(Y=data, x=group)
summary(fit)
head(fit$results$final)
}
