% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastANCOM.R
\name{fastANCOM}
\alias{fastANCOM}
\title{fastANCOM, a fast method for analysis of composition of microbiomes}
\usage{
fastANCOM(
  Y,
  x,
  Z = NULL,
  rand = NULL,
  pseudo = 0.5,
  outlier = F,
  sig = 0.05,
  detect.rate = 0.7,
  ref.rate = 0.05
)
}
\arguments{
\item{Y}{the abundance table, both count/relative or any smoothed data matrix}

\item{x}{the variable of interest}

\item{Z}{the other covariates to be adjusted}

\item{rand}{the random variable for the mixed effect model}

\item{pseudo}{the pseudo number for zero smoothing, default is 0.05}

\item{outlier}{whether using cook's distance for outlier detection or not, default is FALSE}

\item{sig}{level of significance to count the reject number, default is 0.05}

\item{detect.rate}{a numerical fraction between 0.5 and 1. Taxa with proportion of the reject number greater than it will be judged as be associated with the variable of interest, i.e., x, default is 0.7}

\item{ref.rate}{a numerical fraction between 0 and 0.5 which means taking these microbes with the least reject proportion as reference to estimate the effect size in ecosystem, default is 0.05}
}
\value{
a dataframe with 5 columns,
\itemize{
\item log2FC, the adjusted effect size from log-linear model
\item log2FC.SD, the standard errors (SEs) of log2FC
\item log2FC.pval, the p-value from Wald test
\item Reject.number, the reject number for each microbe
\item REJECT, the final result whether each microbe to be associated with the variable of interest or not
}
}
\description{
The framework of log-linear regression is used for the log transformed data and
then build hypothesis with the variable of interest for inference at the ecosystem level with the framework of ANCOM.
fastANCOM fits the mariginal model and then computing the effect for the joint model
}
\examples{
data <- matrix(rpois(100*60, 10), 60)
b <- rnorm(10, 4)
data[31:60,1:10] <- t(t(data[31:60,1:10])*2^b)
group <- rep(c(0, 1), each=30)
fit <- fastANCOM(Y=data, x=group)
head(fit)
}
